Move all data from deployment to operator only
