+++ src/containers/DeploymentFlowContainer/DeploymentFlowContainer.jsx
@@ -10,8 +10,8 @@ import {
   saveDeploymentOperatorPosition,
 } from 'store/operator';
 
-const operatorsSelector = ({ deploymentOperatorsReducer }) => {
-  return deploymentOperatorsReducer;
+const operatorsSelector = ({ operatorsReducer }) => {
+  return operatorsReducer;
 };
 
 const loadingSelector = ({ uiReducer }) => {
