--- src/store/deployments/deploymentOperator/actionTypes.js
@@ -1,30 +1,0 @@
-// ACTION TYPES
-const CREATE_DEPLOYMENT_OPERATOR_REQUEST = 'CREATE_DEPLOYMENT_OPERATOR_REQUEST';
-const CREATE_DEPLOYMENT_OPERATOR_SUCCESS = 'CREATE_DEPLOYMENT_OPERATOR_SUCCESS';
-const CREATE_DEPLOYMENT_OPERATOR_FAIL = 'CREATE_DEPLOYMENT_OPERATOR_FAIL';
-
-const REMOVE_DEPLOYMENT_OPERATOR_REQUEST = 'REMOVE_DEPLOYMENT_OPERATOR_REQUEST';
-const REMOVE_DEPLOYMENT_OPERATOR_SUCCESS = 'REMOVE_DEPLOYMENT_OPERATOR_SUCCESS';
-const REMOVE_DEPLOYMENT_OPERATOR_FAIL = 'REMOVE_DEPLOYMENT_OPERATOR_FAIL';
-
-const SET_DEPLOYMENT_OPERATOR_PARAMETERS_REQUEST =
-  'SET_DEPLOYMENT_OPERATOR_PARAMETERS_REQUEST';
-const SET_DEPLOYMENT_OPERATOR_PARAMETERS_SUCCESS =
-  'SET_DEPLOYMENT_OPERATOR_PARAMETERS_SUCCESS';
-const SET_DEPLOYMENT_OPERATOR_PARAMETERS_FAIL =
-  'SET_DEPLOYMENT_OPERATOR_PARAMETERS_FAIL';
-
-// EXPORT
-export default {
-  CREATE_DEPLOYMENT_OPERATOR_REQUEST,
-  CREATE_DEPLOYMENT_OPERATOR_SUCCESS,
-  CREATE_DEPLOYMENT_OPERATOR_FAIL,
-
-  REMOVE_DEPLOYMENT_OPERATOR_REQUEST,
-  REMOVE_DEPLOYMENT_OPERATOR_SUCCESS,
-  REMOVE_DEPLOYMENT_OPERATOR_FAIL,
-
-  SET_DEPLOYMENT_OPERATOR_PARAMETERS_REQUEST,
-  SET_DEPLOYMENT_OPERATOR_PARAMETERS_SUCCESS,
-  SET_DEPLOYMENT_OPERATOR_PARAMETERS_FAIL,
-};
