--- src/store/deployments/deploymentOperator/index.js
@@ -1,7 +1,0 @@
-import Actions from './actions';
-import ActionTypes from './actionTypes';
-import Reducer from './deploymentOperatorReducer';
-
-export default Reducer;
-
-export { Actions, ActionTypes };
