--- src/store/deployments/deploymentOperators/actionTypes.js
@@ -1,17 +1,0 @@
-const FETCH_DEPLOYMENT_OPERATORS_REQUEST =
-  '@DEPLOYMENT_OPERATORS/FETCH_DEPLOYMENT_OPERATORS_REQUEST';
-const FETCH_DEPLOYMENT_OPERATORS_SUCCESS =
-  '@DEPLOYMENT_OPERATORS/FETCH_DEPLOYMENT_OPERATORS_SUCCESS';
-const FETCH_DEPLOYMENT_OPERATORS_FAIL =
-  '@DEPLOYMENT_OPERATORS/FETCH_DEPLOYMENT_OPERATORS_FAIL';
-
-const CLEAR_ALL_DEPLOYMENT_OPERATORS =
-  '@DEPLOYMENT_OPERATORS/CLEAR_ALL_DEPLOYMENT_OPERATORS';
-
-export default {
-  FETCH_DEPLOYMENT_OPERATORS_REQUEST,
-  FETCH_DEPLOYMENT_OPERATORS_SUCCESS,
-  FETCH_DEPLOYMENT_OPERATORS_FAIL,
-
-  CLEAR_ALL_DEPLOYMENT_OPERATORS,
-};
