--- src/store/deployments/deploymentOperators/index.js
@@ -1,7 +1,0 @@
-import Actions from './actions';
-import ActionTypes from './actionTypes';
-import Reducer from './deploymentOperatorsReducer';
-
-export default Reducer;
-
-export { Actions, ActionTypes };
