+++ src/store/index.js
@@ -3,8 +3,6 @@ import thunk from 'redux-thunk';
 
 import compareResultsReducer from './compareResults/compareResultsReducer';
 import deploymentsReducer from './deployments/deploymentsReducer';
-import deploymentOperatorReducer from './deployments/deploymentOperator/deploymentOperatorReducer';
-import deploymentOperatorsReducer from './deployments/deploymentOperators/deploymentOperatorsReducer';
 import deploymentRunsReducer from './deployments/deploymentRuns/deploymentRunsReducer';
 import Projects from './projects';
 import experimentRunReducer from './projects/experiments/experimentRuns/experimentRunsReducer';
@@ -26,8 +24,6 @@ import { monitoringsReducer } from './monitorings';
 const reducers = combineReducers({
   compareResultsReducer,
   deploymentsReducer,
-  deploymentOperatorReducer,
-  deploymentOperatorsReducer,
   deploymentRunsReducer,
   Projects,
   experimentRunReducer,
