+++ src/store/operators/operators.actionTypes.js
@@ -9,3 +9,6 @@ export const UPDATE_OPERATOR_DEPENDENCIES =
   '@OPERATORS/UPDATE_OPERATOR_DEPENDENCIES';
 
 export const UPDATE_OPERATORS_OPTIONS = '@OPERATORS/UPDATE_OPERATORS_OPTIONS';
+
+export const CLEAR_ALL_DEPLOYMENT_OPERATORS =
+  '@OPERATORS/CLEAR_ALL_DEPLOYMENT_OPERATORS';
